+++ build-support/setup-venv.sh
@@ -6,7 +6,7 @@ set -e
 # see https://vsupalov.com/developing-with-python3-8-on-ubuntu-18-04/
 # also don't forget to run sudo apt install python3.9-venv python3.9-dev (otherwise ensurepip module won't be available)
 
-PYTHON_BIN=python3.10
+PYTHON_BIN=python3.11
 VIRTUALENV=build-support/.venv
 PIP="${VIRTUALENV}/bin/pip"
 REQUIREMENTS_FILE=3rdparty/requirements.txt