+++ src/python/arcor2_build/source/tests/test_logic.py
@@ -326,7 +326,7 @@ def test_branched_output_2() -> None:
     assert ac6_idx > if_bool_2_res_true_idx
 
 
-@pytest.mark.xfail()
+@pytest.mark.xfail
 def test_alone_if() -> None:
     scene = Scene("s1")
     obj = SceneObject("test_name", "Test")