+++ src/python/arcor2_calibration/robot.py
@@ -4,12 +4,16 @@ import cv2
 import numpy as np
 import open3d as o3d
 from PIL import Image
-from urdfpy import URDF
 
 from arcor2.data.camera import CameraParameters
 from arcor2.data.common import Joint, Pose
 from arcor2.exceptions import Arcor2Exception
 from arcor2.logging import get_logger
+
+np.int = int  # type: ignore  # TODO workaround for urdfpy
+np.float = float  # type: ignore  # TODO workaround for urdfpy
+
+from urdfpy import URDF  # noqa
 
 logger = get_logger(__name__)