+++ src/python/arcor2_kinect_azure/scripts/kinect_azure.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3.10
+#!/usr/bin/env python
 from arcor2_kinect_azure.app import main
 
 if __name__ == "__main__":