+++ src/python/arcor2_kinect_azure/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [0.8.0] - 2024-04-11
+
+### Changed
+
+- Updated dependencies, switched to Python 3.11.
 
 ## [0.7.0] - 2023-07-20